优化位图和块缓存的内存访问，改进文件系统的目录操作，增强 mkdir 和 rm 命令的功能

+++ easy-fs/src/bitmap.rs
@@ -28,25 +28,22 @@ impl Bitmap {
     /// 分配一个块，返回分配的 bit 的编号
     pub fn alloc(&self, block_device: &Arc<dyn BlockDevice>) -> Option<usize> {
         for block_id in 0..self.blocks {
-            let pos = get_block_cache(
-                block_id + self.start_block_id as usize,
-                Arc::clone(block_device),
-            )
-            .lock()
-            .modify(0, |bitmap_block: &mut BitmapBlock| {
-                if let Some((bits64_pos, inner_pos)) = bitmap_block
-                    .iter()
-                    .enumerate()
-                    .find(|(_, bits64)| **bits64 != u64::MAX)
-                    .map(|(bits64_pos, bits64)| (bits64_pos, bits64.trailing_ones() as usize))
-                {
-                    // modify cache
-                    bitmap_block[bits64_pos] |= 1u64 << inner_pos;
-                    Some(block_id * BLOCK_BITS + bits64_pos * 64 + inner_pos as usize)
-                } else {
-                    None
-                }
-            });
+            let pos = get_block_cache(block_id + self.start_block_id, Arc::clone(block_device))
+                .lock()
+                .modify(0, |bitmap_block: &mut BitmapBlock| {
+                    if let Some((bits64_pos, inner_pos)) = bitmap_block
+                        .iter()
+                        .enumerate()
+                        .find(|(_, bits64)| **bits64 != u64::MAX)
+                        .map(|(bits64_pos, bits64)| (bits64_pos, bits64.trailing_ones() as usize))
+                    {
+                        // modify cache
+                        bitmap_block[bits64_pos] |= 1u64 << inner_pos;
+                        Some(block_id * BLOCK_BITS + bits64_pos * 64 + inner_pos)
+                    } else {
+                        None
+                    }
+                });
             if pos.is_some() {
                 return pos;
             }
